About this project
This is a project I created out of interest. The origin of this project is from the time when I was an intern and was supposed to create a program with VBA that allows the User to operate on the Excel file with an "Userform". This C# project is supposed to be an attempt at a remake of the original program.

It still has some bugs and some features will be added over time.

Known bugs:
-a few code parts need to be surrounded by a try-catch statement
-after searching for a value, an error message appears during sheet selection

Missing functions:
-The search function should include that you can find the next or previous value by pressing a button.
-Values in the cells can not be edited
-Cells are not displayed true to original (missing color and formation)
-Create copy of original workbook and if user clicks "CANCEL BUTTON" delete workbook and change copy to original name.

Upcoming features:
-Open the last edited file
-Option to add a sheet instead of reloading it
-columns and rows need an ID like in Excel
-Dark Mode switch
-Localization
-UI / UX revision